Change data location to private fork for now
I forked the main data repository so that I could repair a couple
of problems with the data for the state of Utah. It eliminated
the strange aberration on May 19 when there was a data outage on
the state's site. Then hospitalization data for May 22 was
completely wrong. Repaired that manually. Looks much better now.